chore: add many more tricks
currently this is a pretty comprehensive list of tricks at around 140 i believe. still need to go through it with a finer tooth comb as the last 40 or i got AI to help me write as i just fed it my knowledge of the tricks that it was writing. not completly optimal but i need SOME data in for now to verify different app functions